+++ pytket/tests/qasm_test.py
@@ -35,6 +35,7 @@ from pytket.circuit import (
     if_not_bit,
     BitRegister,
     CustomGate,
+    RangePredicateOp,
 )
 from pytket.circuit.decompose_classical import DecomposeClassicalError
 from pytket.circuit.logic_exp import BitWiseOp, create_bit_logic_exp
@@ -953,6 +954,49 @@ if(tk_SCRATCH_BIT[0]==1) h q[0];
 measure q[0] -> c[0];
 """
     )
+
+
+def test_conditional_multi_line_ops() -> None:
+    # https://github.com/CQCL/tket/issues/1491
+    c = Circuit(0, 3)
+    c.add_c_setbits(values=[True, True], args=[Bit(1), Bit(2)], condition=Bit(0))
+    qasm = circuit_to_qasm_str(c, header="hqslib1")
+    assert (
+        qasm
+        == """OPENQASM 2.0;
+include "hqslib1.inc";
+
+creg c[3];
+if(c[0]==1) c[1] = 1;
+if(c[0]==1) c[2] = 1;
+"""
+    )
+    c = Circuit(0, 5)
+    c.add_c_copybits(
+        args_in=[Bit(1), Bit(2)], args_out=[Bit(3), Bit(4)], condition=Bit(0)
+    )
+    qasm = circuit_to_qasm_str(c, header="hqslib1")
+    assert (
+        qasm
+        == """OPENQASM 2.0;
+include "hqslib1.inc";
+
+creg c[5];
+if(c[0]==1) c[3] = c[1];
+if(c[0]==1) c[4] = c[2];
+"""
+    )
+
+
+def test_conditional_range_predicate() -> None:
+    range_predicate = RangePredicateOp(6, 0, 27)
+    c = Circuit(0, 8)
+    c.add_gate(range_predicate, [0, 1, 2, 3, 4, 5, 6], condition=Bit(7))
+    with pytest.raises(Exception) as errorinfo:
+        circuit_to_qasm_str(c, header="hqslib1")
+        assert "Conditional RangePredicate is currently unsupported." in str(
+            errorinfo.value
+        )
 
 
 def test_range_with_maxwidth() -> None: